Redis implementation
Add Aika.Redis project
AikaHistorian and IHistorian now use an Init method instead of trying to do everything in the constructor.
Aika.AspNetCore now uses AikaService to initialise the historian in a background task.
SampleDataGenerator in the sample app is now smarter about when it creates tags, and also works better with persistent back-end stores.